budget command bugfixes

+++ code/budget_view.py
@@ -19,9 +19,8 @@ def run(message, bot):
     try:
         print("here")
         chat_id = message.chat.id
-        if helper.isOverallBudgetAvailable(chat_id):
+        if helper.isOverallBudgetAvailable(chat_id) or helper.isCategoryBudgetAvailable(chat_id):
             display_overall_budget(message, bot)
-        elif helper.isCategoryBudgetAvailable(chat_id):
             display_category_budget(message, bot)
         else:
             raise Exception(
@@ -52,7 +51,10 @@ def display_category_budget(message, bot):
     format suitable for display, and returns the same through the bot to the Telegram UI.
     """
     chat_id = message.chat.id
-    data = helper.getCategoryBudget(chat_id)
-    graphing.viewBudget(data)
-    bot.send_photo(chat_id, photo=open("budget.png", "rb"))
-    os.remove("budget.png")
+    if helper.isCategoryBudgetAvailable(chat_id):
+        data = helper.getCategoryBudget(chat_id)
+        graphing.viewBudget(data)
+        bot.send_photo(chat_id, photo=open("budget.png", "rb"))
+        os.remove("budget.png")
+    else:
+        bot.send_message(chat_id, "You are yet to set your budget for different categories.")
